CTCP-538: Allow percent sign on changes to report form
Aligned the validation on the "What do you need to report?" input with the validation on the "What is the new official customs seal number or mark?" input as both allow the same character set.

Also - updated the regex slightly as the hyphen (-) needs to be at the beginning or end of the character set otherwise it's considered a range identifier.